feat:channel edit
