+++ src/cacheKeys.js
@@ -1,15 +1,11 @@
 const CACHEKEYS = {
-  SESSION_DETAILS: "SESSION_DETAILS",
   FETCH_CHANNELS: "FETCH_CHANNELS",
   FETCH_USERS: "FETCH_USERS",
-  FETCH_USAGE: "FETCH_USAGE",
-  FETCH_BITRATE: "FETCH_BITRATE",
-  FETCH_CHANNEL_HEALTH: "FETCH_CHANNEL_HEALTH",
   SESSION_AUTH_DATA: "SESSION_AUTH_DATA",
-  FETCH_SERVERS: "FETCH_SERVERS",
-  FETCH_VERSION: "FETCH_VERSION",
   FETCH_VIEW_CHANNEL: "FETCH_VIEW_CHANNEL",
   FETCH_VIEWS: "FETCH_VIEWS",
+  FETCH_SERVERS: "FETCH_SERVERS",
+  FETCH_SERVER_NAME: "FETCH_SERVER_NAME",
 };
 
 export default CACHEKEYS;
